Funções concatenadas e funcionando em ordens
