+++ src/index.html
@@ -132,8 +132,8 @@
     <div >
       <div class="containers"><button class="button-filter" id="menu-filter">Filtros Avançados<img  src="img/icons8-filtro-25.png" ></button>
 
-      <select id="browsers" class="button-filter">
-        <option value="">Ordenar por</option>
+      <select id="browsers" class="button-filter">     
+        <option>Ordenar por</option>
         <option value="0">De A a Z</option>
         <option value="1">De Z a A</option>
         <option value="2">Do menor para o maior</option>
